move around compare_paths
