+++ crates/project/Cargo.toml
@@ -69,7 +69,6 @@ snippet_provider.workspace = true
 terminal.workspace = true
 text.workspace = true
 util.workspace = true
-unicase.workspace = true
 which.workspace = true
 
 [dev-dependencies]
